+++ anytone-cps.py
@@ -1,12 +1,5 @@
 
 import sys, argparse, os
-
-### DEBUG
-# os.system('python build-ui.py')
-os.system('python compile.py')
-###
-
-
 from PySide6.QtGui import QIcon
 from PySide6.QtWidgets import QApplication
 from qt_material import apply_stylesheet